vis underkjente attesteringer i saksoversikten
