+++ src/api/featureToggleApi.ts
@@ -3,6 +3,7 @@ import apiClient, { ApiClientResult } from './apiClient';
 export enum FeatureToggle {
     Revurdering = 'supstonad.ufore.revurdering',
     Hotjar = 'supstonad.ufore.hotjar',
+    Klage = 'supstonad.ufore.klage',
 }
 
 export type FeatureToggles = {
